Moving shot history state into gameview
